9.0

+++ addons/report/models/report_paperformat.py
@@ -1,23 +1,5 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    OpenERP, Open Source Management Solution
-#    Copyright (C) 2014-Today OpenERP SA (<http://www.openerp.com>).
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from functools import partial
 
@@ -68,10 +50,10 @@ class report_paperformat(osv.Model):
                                             ('custom', 'Custom')],
                                            'Paper size',
                                            help="Select Proper Paper size"),
-                'margin_top': fields.integer('Top Margin (mm)'),
-                'margin_bottom': fields.integer('Bottom Margin (mm)'),
-                'margin_left': fields.integer('Left Margin (mm)'),
-                'margin_right': fields.integer('Right Margin (mm)'),
+                'margin_top': fields.float('Top Margin (mm)'),
+                'margin_bottom': fields.float('Bottom Margin (mm)'),
+                'margin_left': fields.float('Left Margin (mm)'),
+                'margin_right': fields.float('Right Margin (mm)'),
                 'page_height': fields.integer('Page height (mm)'),
                 'page_width': fields.integer('Page width (mm)'),
                 'orientation': fields.selection([('Landscape', 'Landscape'),
